reunite Value and Serialize instance
